render simple sphere

+++ renderer/bsdf.cpp
@@ -2,7 +2,7 @@
 
 fvec3 BSDF::sample_f(const fvec3 &out, fvec3 &in, float &pdf) {
     int num_bxdfs = static_cast<int>(bxdfs.size());
-    int bxdf_idx = arma::randi(arma::distr_param(0, num_bxdfs - 1));
+    int bxdf_idx = static_cast<int>(arma::randi(arma::distr_param(0, num_bxdfs - 1)));
     fvec3 res = bxdfs[bxdf_idx]->sample_f(out, in, pdf);
     pdf /= num_bxdfs;
     return res;
